Altered many core elements of the game code in order to remove multiple instances of the same object and standardize the location of important game data.
Integrated the Turn class into Map in order to remove overlapping code.
While turns are being taken, the Wagon and Map are stored in the mainScreen.
The game saves the player's Wagon and Map in the GameInterface class when changing screens, such as entering/leaving a shop.
Fixed a bug where the shop didn't update to inform the player of an item's new quantity after purchasing additional items.
After leaving a shop, the game is resumed where last left.